vboot: 4GiB DRAM addressing support
To support 4G DRAM addressing, we apply uint64_t on firmware component
address in coreboot.
Hence, we also need to apply the same thing in depthcharge.
Otherwise, depthcharge will be blocked at the place which checks the
vboot handoff size.

BUG=none
TEST=compiled with Oak board
BRANCH=none

+++ src/vboot/util/vboot_handoff.h
@@ -30,7 +30,7 @@
 #define MAX_PARSED_FW_COMPONENTS 5
 
 struct firmware_component {
-	uint32_t address;
+	uint64_t address;
 	uint32_t size;
 } __attribute__((packed));
